refactor(server): split registry modules
